requirements.txt
requests==2.26.0
SpeechRecognition==3.8.1
python-dotenv==0.19.1